requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
redis==5.2.0
celery==5.4.0
django-celery-beat==2.7.0
Pillow==11.0.0
openpyxl==3.1.5
qrcode[pil]==8.0
reportlab==4.2.5
gunicorn==23.0.0
python-decouple==3.8
django-extensions==3.2.3
dj-database-url==2.2.0
django-filter==24.3